lsvpd : Changed rtasgetVpd signature to handle error case
rtasGetVPD has return type as unsigned long int but at the error condition it
sends -ve value, which doesn't make any sense. At error also return value is
stored in unsigned long, so error condition also will be treated as success.
By making return type of rtasGetVPD function as ssize_t, we will be able to
handle error condition.
As ssize_t on 32 bit machine can store the value till "2147483647" which is
"2 GB". By using integer also we are able to handle all vpd info provided by
phyp call.

This patch fixes the above error handling issue by converting return type from
"unsigned long int" -> "ssize_t".

[Fixed return type of other getRtasVPD() - Vasant]

+++ src/include/rtascollector.hpp
@@ -12,6 +12,7 @@
 #include <string>
 #include <vector>
 #include <errno.h>
+#include <stdlib.h>
 
 
 
@@ -60,6 +61,6 @@ namespace lsvpd
 			 * @param data
 			 *   A pointer to the address of the buffer to hold requested VPD
 			 */
-			static unsigned long int rtasGetVPD(const string& yl, char** data);
+			static ssize_t rtasGetVPD(const string& yl, char** data);
 	};
 }
